own test file fro medium
